% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even_or_odd.R
\name{even_or_odd}
\alias{even_or_odd}
\title{Check if a number is even or odd}
\usage{
even_or_odd(x)
}
\arguments{
\item{x}{A whole number (or a vector of whole numbers).}
}
\value{
"even" or "odd" for each number.
}
\description{
Check if a number is even or odd
}
